10. BoardListRespDto에 totalCount 변수 Long에서 int로 자료형 변경

+++ src/main/java/com/book/dto/board/BoardListRespDto.java
@@ -23,9 +23,8 @@ public class BoardListRespDto {
     private Integer prev;
     private Integer next;
     private List<Integer> pageNumbers;
-    private Long totalCount;
+    private int totalCount;
     private Boolean isPrev;
     private Boolean isNext;
-    
     private Boolean isFirst;
 }
